refactor: refactored about section with tech stack into a separate component
