+++ pages/index.tsx
@@ -33,6 +33,7 @@ import Navbar from "@/components/navbar"
 import IntroSection from "@/components/introSection"
 import AboutSection from "@/components/aboutSection"
 import DrDerekAustin from "@/images/DrDerekAustin.png"
+import TechStackSection from "@/components/techStackSection"
 
 const DesktopSections = [
   D0_Intro_Animation,
@@ -59,23 +60,6 @@ const MobileSections = [
   M6_Blog_A,
   M6_Blog_B,
   M7_Contact,
-]
-
-const techStack = [
-  "Node.js",
-  "Typescript",
-  "React",
-  "Next.js",
-  "Tailwind CSS",
-  "Vercel",
-  "Github Actions",
-  "React Query",
-  "Redux / Redux Toolkit",
-  "Jest",
-  "Cypress",
-  "React Testing Library",
-  "Playwright",
-  "Tailwind CSS",
 ]
 
 /** Helper function to join Tailwind CSS classNames. Filters out falsy values */
@@ -127,42 +111,7 @@ function DisplaySections({
 
             {/*========= ABOUT SECTION WITH TECH STACK ========= */}
             <div className="section bg-[#b9e3ff]">
-              <div className="h-screen">
-                <div className="h-2/5">
-                  <div className="w-4/5 mx-auto">
-                    <div className="mb-auto pt-3">
-                      <h2 className="text-6xl">About</h2>
-                    </div>
-                    {/*========= DR DEREK AUSTIN IMAGE ========= */}
-                    <div className="pt-4">
-                      <Image
-                        src={DrDerekAustin}
-                        alt="Dr Derek Austin"
-                        object-fit="contain"
-                      />
-                    </div>
-                  </div>
-                </div>
-
-                {/* ========= TECH STACK ============ */}
-                <div className="h-3/5">
-                  <div className="w-4/5 mx-auto">
-                    <p className="pt-5 pb-3 text-2xl">Tech Stack</p>
-                  </div>
-                  <div className="w-4/5 mx-auto">
-                    <div className="flex flex-wrap gap-x-4 content-between w-5/6">
-                      {techStack.map((item: string) => (
-                        <p
-                          key={item}
-                          className="bg-[#FFE366] rounded-tr-xl pl-2 pr-2 py-1 text-xl mb-2"
-                        >
-                          {item}
-                        </p>
-                      ))}
-                    </div>
-                  </div>
-                </div>
-              </div>
+              <TechStackSection />
             </div>
 
             {/*========= WORK EXPERIENCE SECTION ========= */}
